// FrontEnd/src/app/services/models/refugee-support.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Document } from '../models/document';
import { ProvidedService } from '../models/provided-service';
export interface RefugeeSupport {
  documents?: Array<Document>;
  helpRequest: string;
  question: string;
  reqStatus: 'PENDING' | 'ANSWERED';
  services: Array<ProvidedService>;
  shelterRequest?: string;
  supportId?: number;
}